services/auth_service/requirements.txt
Flask==2.1.3
werkzeug==2.1.2
flask-mongoengine==1.0.0
gunicorn==23.0.0
python-dotenv==1.0.0
PyJWT==2.7.0
requests==2.28.1
psycopg2==2.9.6